Adds a utility module for link-checking, and associated rake task
The LinkChecker module allows for the validating of URLs to ensure
they are active and then extract as much metadata as possible from
them.  Datafiles (the thing that is actually checked) will be saved
back to the database with the new metadata, along with the datetime
that the check was carried out.

Checks can be run from the CLI using rake either at an organisational
level

```
rake linkcheck:organisation[land-registry]
```

or a dataset level

```
rake linkcheck:dataset[1862-act-registers]
```

```
```
```
```