Add basis for authorization.
Uses cancancan (bundle install required) to determine what publishers are
able to do.  The check can be done in each controller with

```
authorize! :update, @dataset
```

and will raise an exception if this is not allowed.

Once edit dataset is done, we can use this to check permissions for ```:create, :update, and :delete```