Adds a temporary page to view metadata quality.
Temporarily adds a new page to allow us to see why we think a dataset
is bad quality, it is at /datasets/NAME/quality.  We may (or not) want
to integrate something similar at some point so that publishers know
why we consider a dataset bad quality.

New scores can be added to the class at /lib/quality/quality_score
by adding a new method that ends with _score, it should return an integer.