Adds a model and a rake task for generating previews.
This is the initial backend of the preview code.

Adds a new model linked to the link table, containing a json blob for
the preview of that link, if it's a CSV.

A new rake task ```preview:generate_all``` will iterate through the database
and find all links claiming to be CSVs (hint they're often not) and trying
to create a preview entry for them.

Because we don't have the faintest idea of what the encoding of the incoming CSV
is, we use the ```file``` command to find out what it is, and convert it to
UTF-8 if possible.

Still needs a UI, changes to dataset view to link to page and hooking in
to new links being created (as a sidekiq task)

Requires a bundle install

Requires a migration