Add theme during import
When importing legacy datasets, this PR will attach a reference to an entry in the Theme
table so that we don't lose data.  We won't be displaying this anywhere right now though.

Some datasets don't have themes, in which case this will be nil.

The default themes are created in db:seed so this will need to be run for themes to be
created.