Ensure syncs are logged correctly
This moves the logic required to perform a legacy sync from the Rake
task to the background worker. This allows us to use the logger
provided by the worker so that we have visibility on what is
happening.

The Rake task now instantiates the worker and performs the work in
process. The worker has been extended to allow the logger to be
specified to support this.