Make database usable after syncing
After downloading the staging database and installing it, we need to set
tell the database that it is a development database (and not production)
as well as deleting an unused function to ensure we can run migrations.

The downside of this is that it's possible that when creating future
migrations, the schema.rb will attempt to include extensions that are
enabled on staging but may/may-not be installed locally.