Remove redundant required variables from schema
The Heroku deployments have their environment set to staging, so we
(only) need to populate the VCAP_SERVICES variable to provide all of the
environment variables.

This is not ideal, and we should look at cleaning up the Heroku
environment variables, as well as improving the setup to include a
throwaway elasticsearch.

But this is just a quick fix to unblock the signon work.