Lint everything
The typical usage across alphagov is to only lint changes that have
been committed to Git. This is typically because projects haven't
gone to the effort of fixing all the historical linting offences.
Accordingly they only want to lint code that has been changed as they
result of a new feature.

We have fixed all our historical offences so it's safe to just lint
everything.