Add custom login page as new home page
The gds-sso helpers require you to call authenticate_user! in order to
check if a user is signed-in. This makes it hard to inject a custom
login page as the check to see if the page should be displayed will
redirect to signon, making the custom login page useless.

A viable workaround is to make the custom login page show if the user
lands on the root path, but not if they land on a specific URL. This
means a new user should still see the custom login page, while an
existing user going to a URL would go straight to signon and login.

Testing logout behaviour with GDS SSO is hard because it forces
automatic login. We should still be testing authentication is required,
though and suggest that GDS SSO test support is improved...
https://github.com/alphagov/gds-sso/issues/82