Do not index fields we don't use in Find
The licence and licence_other fields are not used in Find anymore, they
have been replaced with 4 new fields that provide better coverage of the
licence metadata we require.  These new fields are used during import
and we will require further work to change the fields used in the
Publish UI.  See
https://trello.com/c/mP0NcLrY/93-remove-licence-and-licenceother-fields
for more information.