Set datafile_last_updated_at in Dataset Import Service
This ensures that when a Dataset is imported its `datafile_last_updated_at`
attribute is set. This value is taken from the most recent `last_modified_at`
date for a Datafile belonging to the Dataset, and if that isn't present we use
the most recent `created` date for a Datafile.