Rename Dataset last_updated_at to public_updated_at
We're changing `last_updated_at` to hold the date that a datafile was mostly
recently changed, and in cases where a dataset has no datafiles it will use the
date that metadata was last modified or created. `last_updated_at` is intended
to be used as the public-facing "Last updated" date displayed on search results
pages in Find Data, so we rename it to `public_updated_at` to reflect this
intention.