Update Dataset Import Service to use public_updated_at
We modify the Dataset Import Service to set `public_updated_at` instead of
`last_updated_at`. This attribute will now be set to the most recently modified
datafile's updated date, and in cases where a dataset has no datafiles will be
set to the metadata updated date or created date.