Fix alias updater in dev and make it atomic
The alias updater required an environment variable to specify the
current ES index to replace, which was really annoying to get manually.
Now the current index is determined automatically.

The alias updater used to remove the old alias link and serially and the
new one. A failure at this point would have caused Find to stop working.
Now the alias update (remove old link, add new link) is done atomically.