Choose the correct name for data files
When importing a legacy datafile, the importer was previously just using
the description field.  It should have been checking resource['name']
and only using resource['description'] as a fallback. This resulted in
lots of datasets having links called 'Resource locator' which is a
description of the link, rather than the name.