Remove unused tests for now
We should invest some time to create proper e2e tests in future, but
this approach is unlikely to work with Signon and is currently only
serving to hide the lack of testing. I think we should remove these
tests to make it clear that some new e2e tests are required.