Improve user journeys for editing a dataset
Editing a dataset currently takes a similar journey to the initial
creation wizard, without the ability to cancel when you try to add a
datafile or doc; showing 'Skip this step' in the context of adding a
datafile or doc also doesn't make sense. This is fixed by changing the
dataset page to remove the 'Add Data' link for datafiles and docs. A
user can still add a new datafile or doc via the 'Change' button, which
takes them to a list of all datafiles of docs with a link to add a new
one. Seeing 'Skip this step' in this context makes more sense because of
the multi-click journey to get to the page.

Clicking 'Save and continue' after a adding a datafile currently prompts
the user to add a doc unless the dataset is published, which seems
arbitrary. This is made better by only prompting if there is no doc.