Rake task to reset all of the datasets for a single organisation
As we encounter issues with the synchronisation, we occassionally get
publishers running their harvester and ending up with 1400 datasets that
have not synced to publish, and fall outside the window within which we
check for changes.

Running this task, with the short name of an organisation, and a boolean
will:

* Delete all the datasets for the organisation if the boolean is "true"
* Re-import all of the datasets for the organisation from the legacy API

There is no transaction, if the container runs out of memory half way
through, you need to run it again! This should only be used in dire
emergencies.