Remove the UrlValidator from the Link model
The UrlValidator is problematic for several reasons:

* It occassionally complains about invalid URLs that are perfectly valid
* It uses a HEAD request which may not be supported by the remote server
* It only supports checking HTTP urls (although FTP ones are valid)
* Links are not checked for malicious content

These issues cause the current sync process to break and stop processing
datasets, meaning that we end up with lots of non-synced datasets in
legacy that need to be manually imported (after disabling the
validator).

As the longer term plan is to switch to the GOV.UK link checker, this
feature has been removed until that work can be undertaken.