Increase worker memory limit on staging/production
CloudFoundry uses Docker to launch containers and constrains their heap
memory to a specified number, within the quota of the plan chosen. This
conflicts with Ruby's GC, which cannot be tuned to keep reserved memory
within a limit - it instead works in terms of object counts.

As it's not possible to stop the issue entirely, we can at least reduce
it's frequency by giving the worker a little more memory.