Add stub worker to coordinate full CKAN sync
The aim of this work is to create a list of which datasets need to be
created, updated or deleted by doing a full comparison against the CKAN
API. Everything is namespaced to allow for multiple CKAN API versions

In order to keep the acceptance tests realistic, the worker also
does a minimal amount to get them passing, pending a fuller
implementation that sets all of the appropriate attributes correctly.

Deletion is determined by the absence of a package in the results
returned from CKAN. In order for this to work with manually published
datasets - and noting a user cannot delete a dataset in Publish - the
`legacy_name` field is used to isolate datasets originating from CKAN.

The CKAN client is made up of modules in order to keep the files
readable, but also to make it easier to add the same capabilities to
future clients (for other versions of the CKAN API). The client does not
implement any retry mechanism, as this should be handled naturally by
the job framework that will be executing this code. The client does not
customize any timeouts, as manual testing indicates this is unnecessary.