Remove unused licence and licence_other fields
This fields are replaced with licence_code, licence_title, etc. As prt
of removing these fields, it became clear that the form element to
populate the licence_other field was broken. As the field no longer
exists, it made more sense to remove the redundant code than try and fix
the broken feature, as leaving a reference to licence_other would be
more confusing.