Implement full mapping for dataset attributes
The dataset importer now calls a new mapper service to convert the
package attributes. Becuase of the unhelpful format of package "extras",
the logic to unravel these has been centralised using a new Package
domain model.

Some of the tests still fail pending a change to make certain
validations only apply in the context of manual publishing within this
app.