Remove readonly validation for harvested datasets
Datasets previously had a read_only validation that was in conflict with
the CKAN sync, as this should be able to update harvested datasets.

It would be difficult to use validation contexts to work around this, as
there are many that could apply and only one (sync) that does not.

Because harvested datasets are not shown to the user, it should be OK to
remove this backend protection as it's unlikely to ever be used.