Finish ES integration for manual/sync workflows
The sync workflow now ensures any created or updated datasets are
published, with the tests refactored slightly to avoid too much
duplication from the ones for the manual workflow.

Unpublishing (deletion) of datasets is now supported for manual and
sync workflows. The publishing and unpublishing calls are synchronous,
so that the manual or sync workflow can report the error properly.