Refactor dataset tests to separate publishing
Previously the dataset tests were split over several variously named
files, making it hard to decide where a new feature or change should be
tested. There is now a simple feature-lead convention for naming these
files and the tests are reorganised so they are associated with the
feature under test.

Testing dataset publishing was hard because of the highly customised set
of fields being exported. As there is no longer any issue with exporting
all fields to ES, removing the 'only' clause from the as_indexed_json
method made it easier to add tests that check a dataset is being
correctly published.