Remove legacy code for CKAN package sync
The previous sync mechanism is no longer used and can be safely deleted.
There is still some code that is used for importing legacy
organisations, which is why the legacy namespace still remains.

The import script no longer tries to import packages, as this tends to
take a long time and is usually unnecessary. It is still possible to
manually import packages by running the PackageSyncWorker.