Simplify the elastic config (less code, more YML)
The ES initialization logic was confusing because it was coupled to the
associated YML file. Making the YML file do more work helps to clarify
where the configuration is coming from in each environment.

It should not be necessary to verify the ES certificate in the same way
that we don't verify the identity of the Postgres DB that the app is
connecting to. Not that the ES client automatically discovers nodes.