Ignore missing datasets when unpublishing
It's possible for a dataset to be unpublished but not deleted as these
operations span multiple databases. We can safely catch a NotFound error
when unpublishing because the operation becomes a no-op; this should
only occur when a previous batch delete has failed and does not indicate
a consistency issue in general, provided the sync workers run regularly.