Use the INSPIRE ref date for public_updated_at
The INSPIRE reference dates are a better indicator of recency for a
dataset, where available. If a dataset has INSPIRE metadata, we will use
the maximum of its reference dates as the dataset public_updated_at,
provided these dates are valid (day, month, year).

From user examples we can see that, while 'revision' is commonly the
right reference date to use, sometimes this is not available but another
reference field like 'published' can serve as a suitable indiactor of
recency; to avoid complex logic, we'll start by maxing over them all.