Limit concurrency and retries of sidekiq workers
The issue:
The sync queue for DGU has been found to clog up, with sometimes
millions of import jobs queued.  Expectation of number of import jobs
queued is in the double digits, with 46k being the absolute maximum as
that's the number of datasets.

The cause:
It has been observed that during these blockages, most or all 5
processes are working on sync jobs, with import jobs trickling through
or not at all.  With new sync jobs appearing every 10 minutes and import
jobs taking 3 weeks to expire, this causes a mildly exponential increase
in queued jobs.  Current hypothesis is that the multiple workers
requesting 46k dataset slugs in pages of 1000 each is slowing down CKAN
and causing the backup.  Since only sync jobs are being processed, more
import workers are queued continuously.

This solution:
By limiting sync jobs to 1 concurrent worker we should allow import jobs
to be processed regardless of the number, with sync jobs waiting
patiently for their turn.  Expiring the (transient) import workers after
a few tries means they won't hang around for weeks if there is an issue.