Always update draft datasets if they're present in CKAN
Previously, records that were marked as draft in the local Postgres database would never get marked as published, even if they are made available again though the CKAN API.

This updates that logic so that datasets which are returned in the CKAN API response but are draft in the local database are always updated, thus marking them as published again.

This is particularly useful for recovering from a repeated class of incident where the importer mistakenly deletes (makes draft) several thousand packages.