Ignore validations for links (and others)
The original design for sync required that all imported datasets should
be editable within the new publishing app, which meant running
validations on all the models on save. While this raised no initial
issues, it turns out that only once a datafile has been saved do the
datafile validations start to apply i.e. first save works, second fails.

As the new publish app is no longer a going concern, it's no longer
sensible to enforce its arbitrary validations and instead disable them
where the tests show this is required. This is the case for datafiles,
where a constraint on the transient year attribute prevented updates.