Set staging ckan base url to data.gov.uk
The CKAN staging server has been turned off. Since the staging url no
longer resolves, running staging publish on the Paas now results in
errors being thrown when the Sidkekiq workers try to sync data from the
staging CKAN. Setting this url to the live version of CKAN should stop
this error occurring. This does mean that we will be syncing the live
data to both staging and production versions hosted on the Paas.