Only count deletions against the number of published documents
There have been a large number of CKAN::V26::Depaginator::DeletionTooLargeError
exceptions recently.  This is preventing the sync from CKAN to Publish from
running correctly.  This was happening because we were comparing the total
number of datasets against the new number of datasets, rather than just
comparing the published datasets against the new number of datasets.