Revamp the way we calculate public_updated_at
Previously we calculated the public_updated_at field for a dataset based
on the first available timestamp in the list of things to try. This
changes the way we calculate this field as part of the publish payload,
so that we try to get the latest timestamp out of the ones available.