Script to fix organogram filenames in the database
This script is run to fix existing organogram urls that are not in
sync with CKAN. It's dependendent on https://github.com/alphagov/ckanext-datagovuk/pull/166
which outputs a csv file with the updated urls.
This csv can then be used in this script to find and replace the
affected urls in the database.

There were some validation issues when trying to save records, this currently
has to be set as `false` as explained in https://github.com/alphagov/datagovuk_publish/commit/5cb26b5543235b105c80ed37669eebd1bf7a7680