Ignore 404 errors in non-production environments
In staging, the CKAN postgres database is synced from production every
day. When this was set up we made a decision to not sync Solr because
it's a new technology on GOV.UK, it takes 3 days to reindex and it would
involve writing custom sync code for Solr.

When accessing the CKAN API the `/api/3/search/dataset` endpoint is
backed by the data in Solr which means that it returns results for
datasets which have subsequently been removed (during the data sync
process if the dataset is removed in production). The import worker then
tries to access more information about those datasets but that endpoint
(which is now backed by Postgres) returns a 404 error.

In production this tells us about a real problem in our system (Solr is
out of sync with Postgres) but in staging it's not very useful, and is
filling up our Sentry quota.