Remove acronym inflection for API
This was added in d021903 to namespace certain controllers but this
no longer applies. However it has the unfortunate side-effect of
breaking the gds-sso gem's namespacing for the user controller because
it uses the title-case variant 'Api::UsersController'. When the zeitwerk
autoloader sees the acronym and the api directory it auto generates the
module 'API' but then runs into problems trying to generate the controller
and raises a NameError for the constant 'Api'. We can remove the acronym
as there is no longer anything namespaced under 'API' within this app.