Correctly assign updated date to links
We are currently assigning the `created` date from CKAN to both the
`created_at` and `updated_at` values for links in the frontend of
data.gov.uk.

This leads to a situation where publishers update data, but the updated
date does not change as it remains fixed on the creation date,

Making this change to correctly set the updated date.