Set test date further in the past
The data in the test fixture has:

```
{
    "id": "7509bbb5-ce6a-4801-8d77-b72c58d46180",
    "organization": "aberdeen-city-council",
    "metadata_modified": "2018-03-18T09:03:05.821Z"
},
```

Note that 2018-03-18 is more than 5 years ago (and has been since March
2023). The only thing about the "dataset_to_update" that meant it needed
to be updated was that it was "updated_at: 5.years.ago". Since that
stopped being true, the test has been failing.

The right thing to do would be to set updated_at to a fixed value which
is before the metadata_modified date, but it's more fun to leave this
bug for future DGU developers 1995 years in the future in the year 4018.